+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Note that since we don't clearly distinguish between a public and private interf
     - Add `auto` that picks `scaled` for mobile devices and `native` elsewhere
     - Add `resolution-mode` Viewer GET param
     - Add `PluginConfig.General.ResolutionMode` config item
+- Default to `blended` transparency on iOS due to `wboit` not being supported.
 
 ## [v4.7.0] - 2024-09-29